table_generator: rapsearch -> taxonomy_lookup
